refactor: combine departure-favorites and stops
