+++ src/api/stops/index.ts
@@ -3,9 +3,15 @@ import { IStopsService } from '../../service/interface';
 import {
   DepartureGroupsPayload,
   DepartureGroupsQuery,
-  DepartureRealtimeQuery
+  DepartureRealtimeQuery,
+  DepartureFavoritesPayload,
+  DepartureFavoritesQuery
 } from '../../service/types';
-import { getDepartureRealtime, getDeparturesCursoredRequest } from './schema';
+import {
+  getDepartureRealtime,
+  getDeparturesCursoredRequest,
+  getDepartureFavoritesCursoredRequest
+} from './schema';
 
 export default (server: Hapi.Server) => (service: IStopsService) => {
   server.route({
@@ -20,6 +26,21 @@ export default (server: Hapi.Server) => (service: IStopsService) => {
       const location = request.payload as unknown as DepartureGroupsPayload;
       const query = request.query as unknown as DepartureGroupsQuery;
       return (await service.getDeparturesGrouped(location, query)).unwrap();
+    }
+  });
+  server.route({
+    method: 'POST',
+    path: '/bff/v2/departure-favorites',
+    options: {
+      tags: ['api', 'stops', 'departures'],
+      validate: getDepartureFavoritesCursoredRequest,
+      description: 'Get departures grouped on lines from feature location'
+    },
+    handler: async (request, h) => {
+      const location = request.payload as unknown as DepartureFavoritesPayload;
+      const query = request.query as unknown as DepartureFavoritesQuery;
+
+      return (await service.getDeparturesFavorites(location, query)).unwrap();
     }
   });
   server.route({
