+++ src/service/impl/stops/journey-gql/departure-group.graphql
