+++ src/service/impl/stops/journey-gql/departure-time.graphql-gen.ts
