+++ src/service/impl/stops/utils/grouping.ts
@@ -6,7 +6,7 @@ import {
   TransportSubmode
 } from '../../../../graphql/journey/journeyplanner-types_v3';
 import { FavoriteDeparture } from '../../../types';
-import { GroupsByIdQuery } from '../../departure-favorites/journey-gql/departure-group.graphql-gen';
+import { GroupsByIdQuery } from '../journey-gql/departure-group.graphql-gen';
 
 type Notice = { text?: string };
 type Situation = {
